WebChat client sends only the text of message
