+++ src/main/java/cz/muni/fi/pv243/spatialtracker/webchat/websocket/WebChatWebSocket.java
@@ -11,7 +11,7 @@ import javax.websocket.Session;
 import javax.websocket.server.PathParam;
 import javax.websocket.server.ServerEndpoint;
 
-import cz.muni.fi.pv243.spatialtracker.webchat.model.NewWebChatMessage;
+import cz.muni.fi.pv243.spatialtracker.webchat.model.NewWebChatMessageEvent;
 import cz.muni.fi.pv243.spatialtracker.webchat.store.KeySessionStore;
 
 @ServerEndpoint(value = "/socket/messages/{room}", encoders = { WebChatMessageEncoder.class })
@@ -32,7 +32,7 @@ public class WebChatWebSocket {
 		sessions.removeSession(roomName, session);
 	}
 
-	public void updateClients(@Observes NewWebChatMessage newMessage) {
+	public void updateClients(@Observes NewWebChatMessageEvent newMessage) {
 		String roomName = newMessage.getRoomName();
 		if (roomName == null) {
 			throw new NullPointerException("roomName");
